+++ src/typescript-worker/build.ts
@@ -5,27 +5,31 @@
  */
 
 import {BareModuleTransformer} from './bare-module-transformer.js';
-import {ImportMapResolver} from './import-map-resolver.js';
 
 import type {
   SampleFile,
   BuildOutput,
-  TypeScriptWorkerConfig,
+  WorkerConfig,
 } from '../shared/worker-api.js';
-import {CachingCdn} from './caching-cdn.js';
-import {getBuilder} from './project-cache.js';
+import {getWorkerContext} from './worker-context.js';
+import {processTypeScriptFiles} from './typescript-builder.js';
 
 export const build = async (
   files: Array<SampleFile>,
-  config: TypeScriptWorkerConfig,
+  config: WorkerConfig,
   emit: (result: BuildOutput) => void
 ): Promise<void> => {
-  const moduleResolver = new ImportMapResolver(config.importMap);
-  const cdn = new CachingCdn(config.cdnBaseUrl ?? 'https://unpkg.com/');
-  const tsBuilder = getBuilder(config, cdn, moduleResolver);
-  const bareModuleBuilder = new BareModuleTransformer(cdn, moduleResolver);
+  const workerContext = getWorkerContext(config);
+  const bareModuleBuilder = new BareModuleTransformer(
+    workerContext.cdn,
+    workerContext.importMapResolver
+  );
+
   const results = bareModuleBuilder.process(
-    tsBuilder.process(files.map((file) => ({kind: 'file', file})))
+    processTypeScriptFiles(
+      workerContext,
+      files.map((file) => ({kind: 'file', file}))
+    )
   );
   for await (const result of results) {
     emit(result);