+++ src/typescript-worker/diagnostic.ts
@@ -1,0 +1,52 @@
+/**
+ * @license
+ * Copyright 2021 Google LLC
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+import ts from '../internal/typescript.js';
+import type * as lsp from 'vscode-languageserver';
+
+/**
+ * Convert a diagnostic from TypeScript format to Language Server Protocol
+ * format.
+ */
+export function makeLspDiagnostic(tsDiagnostic: ts.Diagnostic): lsp.Diagnostic {
+  return {
+    code: tsDiagnostic.code,
+    source: tsDiagnostic.source ?? 'typescript',
+    message: ts.flattenDiagnosticMessageText(tsDiagnostic.messageText, '\n'),
+    severity: diagnosticCategoryMapping[tsDiagnostic.category],
+    range: {
+      start:
+        tsDiagnostic.file !== undefined && tsDiagnostic.start !== undefined
+          ? tsDiagnostic.file.getLineAndCharacterOfPosition(tsDiagnostic.start)
+          : {character: 0, line: 0},
+      end:
+        tsDiagnostic.file !== undefined &&
+        tsDiagnostic.start !== undefined &&
+        tsDiagnostic.length !== undefined
+          ? tsDiagnostic.file.getLineAndCharacterOfPosition(
+              tsDiagnostic.start + tsDiagnostic.length
+            )
+          : {character: 0, line: 0},
+    },
+  };
+}
+
+/**
+ * We don't want a runtime import of 'vscode-languageserver' just for the
+ * DiagnosticSeverity constants. We can duplicate the values instead, and assert
+ * we got them right with a type constraint.
+ */
+const diagnosticCategoryMapping: {
+  [ts.DiagnosticCategory.Error]: typeof lsp.DiagnosticSeverity['Error'];
+  [ts.DiagnosticCategory.Warning]: typeof lsp.DiagnosticSeverity['Warning'];
+  [ts.DiagnosticCategory.Message]: typeof lsp.DiagnosticSeverity['Information'];
+  [ts.DiagnosticCategory.Suggestion]: typeof lsp.DiagnosticSeverity['Hint'];
+} = {
+  [ts.DiagnosticCategory.Error]: 1,
+  [ts.DiagnosticCategory.Warning]: 2,
+  [ts.DiagnosticCategory.Message]: 3,
+  [ts.DiagnosticCategory.Suggestion]: 4,
+};